contrib/convert-opus-mt/requirements.txt
protobuf==4.24.3
PyYAML==6.0.1
